+++ packages/cli/__tests__/zosfiles/__integration__/download/am/cli.files.download.am.integration.test.ts
@@ -47,4 +47,13 @@ describe("Download All Member", () => {
         expect(response.stderr.toString()).toContain("Missing Positional");
     });
 
+    it("should fail due to specifying both binary and record", async () => {
+        const shellScript = path.join(__dirname, "__scripts__", "command", "command_download_all_member.sh");
+        const response = runCliScript(shellScript, TEST_ENVIRONMENT, ["test", "--binary", "--record"]);
+        expect(response.status).toBe(1);
+        expect(response.stderr.toString()).toContain("following options conflict");
+        expect(response.stderr.toString()).toContain("--record");
+        expect(response.stderr.toString()).toContain("--binary");
+    });
+
 });