+++ packages/cli/__tests__/zosfiles/__integration__/upload/ftds/__scripts__/command/command_upload_ftds.sh
@@ -2,7 +2,7 @@
 set -e
 
 echo "================Z/OS FILES UPLOAD DATA SET==============="
-zowe zos-files upload file-to-data-set "$1" $2 --host fakehost --user fakeuser --pw fakepass
+zowe zos-files upload file-to-data-set "$1" $2 --host fakehost --user fakeuser --pw fakepass $3 $4
 if [ $? -gt 0 ]
 then
     exit $?