+++ packages/cli/src/zosfiles/download/ds/Dataset.definition.ts
@@ -43,6 +43,7 @@ export const DatasetDefinition: ICommandDefinition = {
         DownloadOptions.file,
         DownloadOptions.extension,
         DownloadOptions.binary,
+        DownloadOptions.record,
         DownloadOptions.preserveOriginalLetterCase,
         DownloadOptions.encoding
     ].sort((a, b) => a.name.localeCompare(b.name)),